Moved services models into their own app
