+++ tbx/services/templates/services/service_index_page.html
@@ -1,7 +1,7 @@
 {% extends "torchbox/base.html" %}
 {% load torchbox_tags wagtailcore_tags wagtailimages_tags staticfiles %}
 
-{% block content %}    
+{% block content %}
 
     <style>
         .about-hero {
